+++ python/GafferSceneUI/ShaderUI.py
@@ -219,7 +219,7 @@ class _ShaderNamePlugValueWidget( GafferUI.PlugValueWidget ) :
 
 	def __reloadButtonClicked( self, button ) :
 
-		with Gaffer.UndoScope( next( iter( self.getPlugs() ) ).ancestor( Gaffer.ScriptNode ) ) :
+		with Gaffer.UndoScope( self.scriptNode() ) :
 			for plug in self.getPlugs() :
 				plug.node().reloadShader()